Conflicts:
	src/DynamoCore/Core/DynamoController.cs
	src/DynamoCore/Core/dynSettings.cs
	src/DynamoCore/Models/DynamoModel.cs
	src/DynamoCore/Search/SearchElements/DSFunctionNodeSearchElement.cs
	src/DynamoCore/Search/SearchElements/SearchElementBase.cs
	src/DynamoCore/UI/Commands/RecordableCommands.cs
	src/DynamoCore/UI/Views/DynamoView.xaml.cs
	src/DynamoCore/ViewModels/SearchViewModel.cs

+++ src/Engine/ProtoCore/Lang/FunctionPointerEvaluator.cs
@@ -44,7 +44,7 @@ namespace ProtoCore.Lang
 
             int classScopeCaller = (int)stackFrame.GetAt(StackFrame.AbsoluteIndex.kClass).opdata;
             int returnAddr = (int)stackFrame.GetAt(StackFrame.AbsoluteIndex.kReturnAddress).opdata;
-            int blockDecl = (int)procNode.runtimeIndex;
+            int blockDecl = procNode.runtimeIndex;
             int blockCaller = (int)stackFrame.GetAt(StackFrame.AbsoluteIndex.kFunctionCallerBlock).opdata;
             int framePointer = runtimeCore.Rmem.FramePointer;
             StackValue thisPtr = StackValue.BuildPointer(-1);
